removed abstractions from auth context
